refactor(extensions): extract sandbox runtime request helpers
